+++ src/types/mdn.ts
@@ -1,0 +1,4 @@
+export interface MDNIndexEntry {
+	title: string;
+	url: string;
+}
